booking/requirements.txt
ariadne==0.26.2
Flask==3.1.2
requests==2.32.5
graphql-core==3.2.5
requests==2.32.5
dotenv
pymongo==4.15.3
grpcio==1.75.0
grpcio-tools==1.75.0
